+++ mindcode/src/main/java/info/teksol/mindcode/compiler/optimization/DeadCodeEliminator.java
@@ -77,8 +77,8 @@ class DeadCodeEliminator extends BaseOptimizer {
         final Set<LogicVariable> uselessWrites = new HashSet<>(writes.keySet());
         uselessWrites.removeAll(reads);
         for (LogicVariable key : uselessWrites) {
-            // Preserve global and main variable assignments unless aggressive
-            if ((key.isGlobalVariable() || key.isMainVariable()) && !aggressive()) continue;
+            // Preserve global and main variable assignments unless advanced
+            if ((key.isGlobalVariable() || key.isMainVariable()) && !advanced()) continue;
 
             // Instruction with at most one output argument are removed immediately
             // Other instructions are inspected further to find out they're fully unused