+++ mindcode/src/main/java/info/teksol/mindcode/compiler/optimization/OptimizationLevel.java
@@ -1,17 +1,25 @@
 package info.teksol.mindcode.compiler.optimization;
 
+import java.util.HashMap;
 import java.util.Map;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public enum OptimizationLevel {
-    OFF,
+    NONE,
     BASIC,
-    AGGRESSIVE,
+    ADVANCED,
     ;
 
-    private static final Map<String, OptimizationLevel> VALUE_MAP =
-            Stream.of(values()).collect(Collectors.toMap(e -> e.name().toLowerCase(), e -> e));
+    private static final Map<String, OptimizationLevel> VALUE_MAP = createValueMap();
+
+    private static Map<String, OptimizationLevel> createValueMap() {
+        Map<String, OptimizationLevel> map = new HashMap<>();
+        for (OptimizationLevel level : values()) {
+            map.put(level.name().toLowerCase(), level);
+        }
+        map.put("off", OptimizationLevel.NONE);
+        map.put("aggressive", OptimizationLevel.ADVANCED);
+        return Map.copyOf(map);
+    }
 
     public static OptimizationLevel byName(String level) {
         return VALUE_MAP.get(level.toLowerCase());